# Add a new `Func::wrap_cabi` API