… tests/all/func.rs