@alexcrichton 2021-09-14
Ok I apologize for all the runarounds here, but finishing out this implementation for host->wasm calls being optimized in the C API I've settled on https://github.com/bytecodealliance/wasmtime/pull/3350 which is both a bit faster than this implementation but also takes a different approach, so I'm going to close this.